learning kivy and making my own interactive ui oof
dont use Builder.load_file it will only work inside an App context, wrestled this for too long.
in the end i prefer to just keep the .kv inside the .py as a string. more portalble.
avoiding using App because i would assume it would be harder to make reusable widgets that dont come with unnecessary wrappings or nestings or stuff.